+++ packages/backend/src/scraper/openapi/openapi-scraper.service.ts
@@ -7,7 +7,8 @@ import { Cron } from '@nestjs/schedule';
 import { Stock } from '@/stock/domain/stock.entity';
 import { StockPeriod } from '@/stock/domain/stockPeriod';
 import { StockDaily } from '@/stock/domain/stockDaily.entity';
-import { exists } from 'fs';
+import { getPreviousDate, getTodayDate } from './api/openapiUtil.api';
+import { openApiToken } from './api/openapiToken.api';
 
 type ChartData = {
   stck_bsop_date: string;
@@ -27,79 +28,36 @@ type ChartData = {
 
 @Injectable()
 export class OpenapiScraperService {
-  private config: (typeof openApiConfig)[] = [];
   public constructor(
     private readonly datasourse: DataSource,
     @Inject('winston') private readonly logger: Logger,
   ) {
-    const accounts = openApiConfig.STOCK_ACCOUNT!.split(',');
-    const api_keys = openApiConfig.STOCK_API_KEY!.split(',');
-    const api_passwords = openApiConfig.STOCK_API_PASSWORD!.split(',');
-    if (
-      accounts.length === 0 ||
-      accounts.length !== api_keys.length ||
-      api_passwords.length !== api_keys.length
-    ) {
-      this.logger.warn('Open API Config Error');
-    }
-    for (let i = 0; i < accounts.length; i++) {
-      this.config.push({
-        STOCK_URL: openApiConfig.STOCK_URL,
-        STOCK_ACCOUNT: accounts[i],
-        STOCK_API_KEY: api_keys[i],
-        STOCK_API_PASSWORD: api_passwords[i],
-      });
-    }
-
-    //if (process.env.NODE_ENV === 'production') {
-    //  this.initAccessToken();
-    //  this.initWebSocketKey();
-    //  this.getItemChartPriceCheck();
-    //}
-    this.initAuthenValue();
-  }
-
-  private async initAuthenValue() {
-    await this.initAccessToken();
-    await this.initWebSocketKey();
-    this.getItemChartPriceCheck();
-  }
-
-  @Cron('30 0 * * 1-5')
-  private async initAccessToken() {
-    const updatedConfig = await Promise.all(
-      this.config.map(async (val) => {
-        val.STOCK_API_TOKEN = await this.getToken(val)!;
-        return val;
-      }),
-    );
-    this.config = updatedConfig;
-    console.log(updatedConfig);
-  }
-
-  @Cron('30 0 * * 1-5')
-  private async initWebSocketKey() {
-    this.config.forEach(async (val) => {
-      val.STOCK_WEBSOCKET_KEY = await this.getWebSocketKey(val)!;
-    });
   }
 
   @Cron('0 1 * * 1-5')
   private async getItemChartPriceCheck() {
     const entityManager = this.datasourse.manager;
     const stocks = await entityManager.find(Stock);
-    const configCount = this.config.length;
+    const configCount = openApiToken.configs.length;
     const chunkSize = Math.ceil(stocks.length / configCount);
 
     for (let i = 0; i < configCount; i++) {
       const chunk = stocks.slice(i * chunkSize, (i + 1) * chunkSize);
-      const config = this.config[i];
+      const config = openApiToken.configs[i];
+      this.getChartData(chunk, config);
+    }
+  }
 
-      for await (const stock of chunk) {
-        let endDate = this.getTodayDate();
-        let startDate = this.getPreviousDate(endDate, 3);
-        let hasData = true;
+  private async getChartData(chunk: Stock[], config: typeof openApiConfig) {
+    const baseTime = 1000;
+    let time = 0;
+    for (const stock of chunk) {
+      time += baseTime;
+      let endDate = getTodayDate();
+      let startDate = getPreviousDate(endDate, 3);
+      let hasData = true;
 
+      setTimeout(async (startDate, endDate) => {
         while (hasData) {
           const query = this.getItemChartPriceQuery(
             stock.id!,
@@ -108,71 +66,21 @@ export class OpenapiScraperService {
             'D',
           );
           const data = await this.getItemChartPrice(config, query);
-
-          if (data && data.output2 && data.output2.length > 0) {
+          if (
+            data &&
+            data.output2 &&
+            data.output2.length > 0 &&
+            data.output2[0].stck_bsop_date
+          ) {
             await this.saveChartData(StockDaily, stock.id!, data.output2);
-            endDate = this.getPreviousDate(startDate, 3);
-            startDate = this.getPreviousDate(endDate, 3);
+            endDate = getPreviousDate(startDate, 3);
+            startDate = getPreviousDate(endDate, 3);
           } else {
             hasData = false;
           }
         }
-      }
+      }, time);
     }
-  }
-
-  private async postOpenApi(url: string, body: {}): Promise<any> {
-    try {
-      const response = await axios.post(url, body);
-      return response.data;
-    } catch (error) {
-      throw new Error(`Request failed: ${error}`);
-    }
-  }
-
-  private async getToken(config: typeof openApiConfig): Promise<string> {
-    const body = {
-      grant_type: 'client_credentials',
-      appkey: config.STOCK_API_KEY,
-      appsecret: config.STOCK_API_PASSWORD,
-    };
-    const tmp = await this.postOpenApi(
-      config.STOCK_URL + '/oauth2/tokenP',
-      body,
-    );
-    if (!tmp.access_token) {
-      throw new NotFoundException('Access Token Failed');
-    }
-    return tmp.access_token as string;
-  }
-
-  private async getWebSocketKey(config: typeof openApiConfig): Promise<string> {
-    const body = {
-      grant_type: 'client_credentials',
-      appkey: config.STOCK_API_KEY,
-      secretkey: config.STOCK_API_PASSWORD,
-    };
-    const tmp = await this.postOpenApi(
-      config.STOCK_URL + '/oauth2/Approval',
-      body,
-    );
-    if (!tmp.approval_key) {
-      throw new NotFoundException('WebSocket Key Failed');
-    }
-    return tmp.approval_key as string;
-  }
-
-  private getTodayDate(): string {
-    const today = new Date();
-    return today.toISOString().split('T')[0].replace(/-/g, '');
-  }
-
-  private getPreviousDate(date: string, months: number): string {
-    const currentDate = new Date(
-      date.slice(0, 4) + '-' + date.slice(4, 6) + '-' + date.slice(6, 8),
-    );
-    currentDate.setMonth(currentDate.getMonth() - months);
-    return currentDate.toISOString().split('T')[0].replace(/-/g, '');
   }
 
   private async existsChartData(
@@ -183,7 +91,7 @@ export class OpenapiScraperService {
     return await manager.findOne(entity, {
       where: {
         stock: { id: stock.stock.id },
-        start_time: stock.start_time,
+        created_at: stock.start_time,
       },
     });
   }
@@ -204,9 +112,16 @@ export class OpenapiScraperService {
     data: ChartData[],
   ) {
     for (const item of data) {
+      if (!item || !item.stck_bsop_date) {
+        continue;
+      }
       const stockPeriod = new StockPeriod();
       stockPeriod.stock = { id: stockId } as Stock;
-      stockPeriod.start_time = new Date(item.stck_bsop_date);
+      stockPeriod.start_time = new Date(
+        parseInt(item.stck_bsop_date.slice(0, 4)),
+        parseInt(item.stck_bsop_date.slice(4, 6)),
+        parseInt(item.stck_bsop_date.slice(6, 8)),
+      );
       stockPeriod.close = parseInt(item.stck_clpr);
       stockPeriod.open = parseInt(item.stck_oprc);
       stockPeriod.high = parseInt(item.stck_hgpr);
@@ -222,7 +137,6 @@ export class OpenapiScraperService {
     query: any,
   ): Promise<any> {
     try {
-      console.log;
       const response = await axios.get(
         config.STOCK_URL +
           '/uapi/domestic-stock/v1/quotations/inquire-daily-itemchartprice',
