--- packages/backend/src/scraper/openapi/openapi-scraper.ts
@@ -1,4 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class OpenapiScraper {}
